refactor : 일부 설정 이동
