+++ aiku/aiku-gateway/src/main/java/gateway/SpringGatewayApplication.java
@@ -14,14 +14,14 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 @EnableJpaAuditing
 @EntityScan(basePackages = {"common"})
 @SpringBootApplication(scanBasePackages = {"gateway", "common"})
-public class SpringLoginApplication {
+public class SpringGatewayApplication {
     public static void main(String[] args) {
-        SpringApplication.run(SpringLoginApplication.class);
+        SpringApplication.run(SpringGatewayApplication.class);
     }
 
     private final TestBean testBean;
     @Autowired
-    public SpringLoginApplication(TestBean testBean) {
+    public SpringGatewayApplication(TestBean testBean) {
         this.testBean = testBean;
     }
     @PostConstruct
