+++ aiku/aiku-gateway/src/main/resources/application.yml
@@ -4,10 +4,6 @@ spring:
     username: sa
     password:
     driver-class-name: org.h2.Driver
-  data:
-    redis:
-      host: localhost
-      port: 6379
   jpa:
     hibernate:
       ddl-auto: create
@@ -25,20 +21,6 @@ jwt:
 custom:
   password: ${CUSTOM_PASSWORD}
 
-kakao:
-  base_url: https://kauth.kakao.com
-  app_id: ${KAKAO_APP_ID}
-
 logging:
   level:
     root: INFO
-
-cloud:
-  aws:
-    s3:
-      bucket: ${AWS_BUCKET}
-    stack.auto: false
-    region.static: ap-northeast-2
-    credentials:
-      accessKey: ${AWS_ACCESS}
-      secretKey: ${AWS_SECRET}
