+++ aiku/aiku-main/src/main/java/aiku_main/SpringMainApplication.java
@@ -8,10 +8,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Bean;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @EnableJpaAuditing
+@EnableFeignClients
+@EnableCaching
 @EntityScan(basePackages = {"common"})
 @SpringBootApplication(scanBasePackages = {"aiku_main", "common"})
 public class SpringMainApplication {
