+++ aiku/aiku-main/src/test/java/aiku_main/integration_test/MemberRegisterServiceTest.java
@@ -1,21 +1,22 @@
-package login.service;
+package aiku_main.integration_test;
 
+import aiku_main.application_event.publisher.PointChangeEventPublisher;
+import aiku_main.dto.MemberProfileDto;
+import aiku_main.dto.MemberRegisterDto;
+import aiku_main.dto.NicknameExistDto;
+import aiku_main.dto.NicknameExistResDto;
+import aiku_main.exception.MemberNotFoundException;
+import aiku_main.oauth.KakaoOauthHelper;
+import aiku_main.oauth.OauthInfo;
+import aiku_main.repository.EventRepository;
+import aiku_main.repository.MemberReadRepository;
+import aiku_main.repository.MemberRepository;
+import aiku_main.s3.S3ImageProvider;
+import aiku_main.service.MemberRegisterService;
 import common.domain.Status;
 import common.domain.event.RecommendEvent;
 import common.domain.member.*;
 import jakarta.persistence.EntityManager;
-import login.application_event.publisher.PointChangeEventPublisher;
-import login.dto.MemberProfileDto;
-import login.dto.MemberRegisterDto;
-import login.dto.NicknameExistDto;
-import login.dto.NicknameExistResDto;
-import gateway.exception.MemberNotFoundException;
-import login.oauth.KakaoOauthHelper;
-import login.oauth.OauthInfo;
-import login.repository.EventRepository;
-import gateway.repository.MemberReadRepository;
-import login.repository.MemberRepository;
-import login.s3.S3ImageProvider;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
