+++ aiku/aiku-main/src/test/resources/application.yml
